requirements.txt
streamlit
openai==0.28
PyPDF2
python-dotenv
numpy
pandas
